Addressed PR reviews (inferring model type, nan filtering in evaluate, consolidating tf and torch models)
