--- river_dl/RGCN_v1.py
@@ -1,75 +1,0 @@
-import torch
-import torch.nn as nn
-
-# RGCN - paper version
-class RGCN_v1(nn.Module):
-    # Built off of https://towardsdatascience.com/building-a-lstm-by-hand-on-pytorch-59c02a4ec091
-    def __init__(self, input_dim, hidden_dim, adj_matrix, recur_dropout = 0, dropout = 0, DA=False, device='cpu'):
-        """
-        @param input_dim: [int] number input feature
-        @param hidden_dim: [int] hidden size
-        @param adj_matrix: Distance matrix for graph convolution
-        @param recur_dropout: [float]
-        @param dropout: [float]
-        @param DA: [bool] If true, returns h and c states as well as predictions
-        """
-        super().__init__()
-        
-        # New stuff
-        self.A = torch.from_numpy(adj_matrix).float().to(device) # provided at initialization
-        # parameters for mapping graph/spatial data
-        self.weight_q = nn.Parameter(torch.Tensor(hidden_dim, hidden_dim))
-        self.bias_q = nn.Parameter(torch.Tensor(hidden_dim))
-        
-        self.input_dim = input_dim
-        self.hidden_size = hidden_dim
-        self.weight_ih = nn.Parameter(torch.Tensor(input_dim, hidden_dim * 4))
-        self.weight_hh = nn.Parameter(torch.Tensor(hidden_dim, hidden_dim * 4))
-        self.bias = nn.Parameter(torch.Tensor(hidden_dim * 4))
-        self.init_weights()
-        
-        self.dropout = nn.Dropout(dropout)
-        self.recur_dropout = nn.Dropout(recur_dropout)
-        
-        self.dense = nn.Linear(hidden_dim, 1)
-        self.DA = DA
-    
-    def init_weights(self):
-        for p in self.parameters():
-            if p.data.ndimension() >= 2:
-                nn.init.xavier_uniform_(p.data)
-            else:
-                nn.init.zeros_(p.data)
-        
-    def forward(self, x, init_states = None):
-        """Assumes x is of shape (batch, sequence, feature)"""
-        bs, seq_sz, _ = x.size()
-        hidden_seq = []
-        if init_states is None:
-            h_t, c_t = (torch.zeros(bs, self.hidden_size).to(x.device), 
-                        torch.zeros(bs, self.hidden_size).to(x.device))
-        else:
-            h_t, c_t = init_states
-        
-        x = self.dropout(x)
-        HS = self.hidden_size
-        for t in range(seq_sz):
-            x_t = x[:, t, :]
-            # batch the computations into a single matrix multiplication
-            gates = x_t @ self.weight_ih + h_t @ self.weight_hh + self.bias
-            i_t, f_t, g_t, o_t = (
-                torch.sigmoid(gates[:, :HS]), # input
-                torch.sigmoid(gates[:, HS:HS*2]), # forget
-                torch.tanh(gates[:, HS*2:HS*3]),
-                torch.sigmoid(gates[:, HS*3:]), # output
-            )
-            q_t = torch.tanh(h_t @ self.weight_q + self.bias_q)
-            c_t = f_t * (c_t + self.A @ q_t) + i_t * self.recur_dropout(g_t) # note: self.A @ q_t
-            h_t = o_t * torch.tanh(c_t)
-            hidden_seq.append(h_t.unsqueeze(1))
-        hidden_seq = torch.cat(hidden_seq, dim= 1)
-        out = self.dense(hidden_seq)
-        if self.DA:
-            return out, (h_t, c_t)
-        else:
-            return out
